refactor(connectors): ship whatsapp.web in the image
Cloud only runs connector code that ships in its own image: assertCustomConnectorInstallAllowed refuses every organization-supplied artifact under LOBU_CLOUD_MODE, so the connector could never have been installed from examples/. WhatsApp is already first-party surface — whatsapp.ts (Cloud API) ships here — and this is its browser sibling.

The key resolves through the existing convention (whatsapp.web -> whatsapp_web.ts, the same dots-to-underscores rule as google.gmail). Verified with the real compiler: it resolves the key, inlines both the adapter and the helpers, and leaves no unresolved relative import.

The example keeps only what is tenant state: the person-alias rule keyed on connector_key='whatsapp.web'. allowJs moves to the connectors package, where the adapter lives - it is serialised with Function.prototype.toString() and must carry no syntax a transform would rewrite.

Also fixes a vacuous .every(): an empty chatRows recorded a finished backfill that never ran.
