+++ examples/personal-agent/tsconfig.json
@@ -6,7 +6,6 @@
     // Matches the root tsconfig; every connector here imports siblings by path
     // with an explicit .ts extension.
     "allowImportingTsExtensions": true,
-    "allowJs": true,
     "strict": true,
     "skipLibCheck": true,
     "noEmit": true,
